edgex-350-env-injection [WIP]
