+++ options/services.go
@@ -1,0 +1,34 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+
+/*
+ * Copyright (C) 2021 Canonical Ltd
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0'
+ */
+
+package options
+
+import (
+	"github.com/canonical/edgex-snap-hooks/v2/files"
+	"github.com/canonical/edgex-snap-hooks/v2/log"
+)
+
+func isValidService(k string) bool {
+	if files.SnapDataDirectoryExists("config/" + k) {
+		log.Infof("%s is a valid service name", k)
+		return true
+	} else {
+		log.Errorf("%s is an invalid service name", k)
+		return false
+	}
+}
